+++ full-stack-engineer/submissions/sobhanmoadab/public/src/App.js
@@ -1,21 +1,23 @@
-import { Route, Routes, Navigate } from 'react-router-dom'
-import SingleProduct from './components/SingleProduct'
-import ListProduct from './components/ListProduct'
-import Login from './components/Login';
-import Register from './components/Register';
+import { Route, Routes  } from 'react-router-dom'
+import SingleProduct from './components/Product/SingleProduct'
+import ListProduct from './components/Product/ListProduct'
+import Login from './components/Auth/Login';
+import Register from './components/Auth/Register';
+import { PrivateRoute } from './utils/PrivateRoute';
+import { ProtectedRoute } from './utils/ProtectedRoute';
+
 function App() {
-  const userInfo = {
-    userId: localStorage.getItem('id'),
-    token: localStorage.getItem('token')
-  }
   return (
     <>
-      {/* Same as */}
       <Routes>
-        <Route path='/' element={userInfo.token ? <ListProduct /> : <Navigate to='/login' />} />
-        <Route path='/login' element={<Login />} />
-        <Route path='/register' element={<Register />} />
-        <Route path='/:id' element={userInfo.token ? <SingleProduct userInfo={userInfo} /> : <Login />} />
+        <Route path='/' element={<PrivateRoute> <ListProduct /> </PrivateRoute>} />
+
+        <Route path='/:id' element={<PrivateRoute> <SingleProduct /> </PrivateRoute>} />
+
+        <Route path='/login' element={<ProtectedRoute> <Login /> </ProtectedRoute>} />
+        
+        <Route path='/register' element={<ProtectedRoute><Register /></ProtectedRoute>} />
+
       </Routes>
     </>
   );
