+++ full-stack-engineer/submissions/sobhanmoadab/public/src/components/Comment/AddComment.scss
@@ -14,6 +14,7 @@
   width: 100%;
   justify-content: center;
   overflow: hidden;
+  margin-top: -80   px;
   flex-direction: row-reverse;
   height: 150px;
   position: relative;
