--- full-stack-engineer/submissions/sobhanmoadab/public/src/componesnts/components/comment.css
@@ -1,145 +1,0 @@
-@import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css);
-
-fieldset,
-label {
-  margin: 0;
-  padding: 0;
-}
-body {
-  margin: 20px;
-}
-h1 {
-  font-size: 1.5em;
-  margin: 10px;
-}
-
-/****** Style Star Rating Widget *****/
-
-.rating {
-  border: none;
-  float: left;
-}
-
-.rating > input {
-  display: none;
-}
-.rating > label:before {
-  margin: 5px;
-  font-size: 1.25em;
-  font-family: FontAwesome;
-  display: inline-block;
-  content: "\f005";
-}
-
-.rating > label {
-  color: #ddd;
-  float: right;
-}
-
-/***** CSS Magic to Highlight Stars on Hover *****/
-
-.rating > input:checked ~ label, /* show gold star when clicked */
-.rating:not(:checked) > label:hover, /* hover current star */
-.rating:not(:checked) > label:hover ~ label {
-  color: #ffd700;
-} /* hover previous stars in list */
-
-.rating > input:checked + label:hover, /* hover current star when changing rating */
-.rating > input:checked ~ label:hover,
-.rating > label:hover ~ input:checked ~ label, /* lighten current selection */
-.rating > input:checked ~ label:hover ~ label {
-  color: #ffed85;
-}
-
-.rev-section {
-  margin: auto;
-  padding: 0 1rem;
-  max-width: 1100px;
-  text-align: center;
-}
-
-.title {
-  font-size: 2.5rem;
-  text-transform: uppercase;
-  color: rgb(50, 50, 50);
-  margin-bottom: 0.5rem;
-}
-.note {
-  font-size: 1.1rem;
-  color: rgb(150, 150, 150);
-  font-style: italic;
-  color: red;
-}
-
-.reviews {
-  margin: 2rem auto;
-  display: flex;
-  flex-wrap: wrap;
-}
-.review {
-  margin: 1rem 1rem;
-  min-width: 300px;
-  flex: 1;
-}
-
-.body-review {
-  position: relative;
-  background-color: rgb(238, 238, 238);
-  padding: 2.5rem;
-  box-shadow: 2px 2px 10px 3px rgb(225, 225, 225);
-}
-.name-review {
-  font-size: 1.5rem;
-  color: rgb(50, 50, 50);
-  margin-bottom: 0.25rem;
-  margin-top: 1rem;
-  text-align: start;
-  font-weight: bold;
-}
-.place-review {
-  color: red;
-  font-style: italic;
-}
-.rating {
-  color: rgb(253, 180, 42);
-  margin: 1rem 0;
-}
-.desc-review {
-  line-height: 1.5rem;
-  letter-spacing: 1px;
-  color: rgb(150, 150, 150);
-}
-
-@media (max-width: 678px) {
-  .review {
-    margin-top: 1.5rem;
-  }
-}
-.input-wrapper {
-  width: 40%;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.input {
-  width: 100%;
-  border: 2px solid #333;
-  padding: 0.5rem 1rem;
-}
-.body-header {
-  position: absolute;
-  top: 5px;
-  right: 5px;
-}
-
-.action {
-  cursor: pointer;
-}
-
-
-.title {
-  font-size: 1rem;
-  font-weight: bold;
-}
