--- full-stack-engineer/submissions/sobhanmoadab/public/src/componesnts/components/index.js
@@ -1,8 +1,0 @@
-
-import ListProduct from './ListProduct'
-import SingleProduct from './SingleProduct'
-
-export {
-    ListProduct,
-    SingleProduct
-}
